refactor(messages): move shared attachments methods to AttachmentMixin
